exposed action.responded to authorize() function
responded was a private variable used for determining if the authorize()
function had called action.accept() or action.reject().

responded has been made a public attribute of the action object so the
authorize function can check to see whether it has called accept() or
reject(). This is useful for complex control flows.