@aarondruck:
I linked the steps to the slides. I also added a red border rule for the active one that obviously needs to be re-styled. We also need to hint that the steps are linked to the current slide. 

Should we add next & back buttons? I don't want it to auto-cycle.

Also of note, I had to fix indentation stuff, is your editor still set to use spaces instead of tabs?