Implement OT synchronization of text fields in forms.
Uses the server echo feature to define a canonical serialization of all
changes.  If a client receives a change from a peer which serializes
before the changes they have queued, the OT delta transpose functions
are used to reorder the patches, and then the local changes are
retransmitted.

We also use the OT mechanism to handle updates to the cursor position
and selection.