Compatability with marked 0.2.10 and docco 0.6.3.
The <h1> tags in marked 0.2.10 contain the `id` attributes which we were
manually adding in the grunt `addHeaderIds` function, causing the match in
`parseMarkdownOutput` to fail (and grunt to crash).  Allow <h1> tags to
contain (simple) attributes.

In docco 0.6.3, the `docco.parse` function added an optional `config`
parameter, but docco crashes if you use the default value.  Filed and
patched upstream in https://github.com/jashkenas/docco/pull/291 -- but
work around this for now by passing `{languages:{}}` as the third
parameter (which is safely ignored in docco 0.6.2).