getContext("2d")!, ! is ok because the first call to getContext can't fail if it's "2d"
From https://developer.mozilla.org/en-US/docs/Web/API/HTMLCanvasElement/getContext

"Later calls to this method on the same canvas element, with the same contextType argument, will always return the same drawing context instance as was returned the first time the method was invoked. It is not possible to get a different drawing context object on a given canvas element."
And "2d" context is always implemented, contrary to "webgl" for example.